setup GitHub Actions and adjust cmake invocation especially on Windows (Claude)

+++ kotlin/uapmd-binding/src/wasmJsMain/kotlin/dev/atsushieno/uapmd/WasmJsEngine.kt
@@ -104,7 +104,7 @@ class WasmJsSequencerEngine internal constructor(
         val mod = wasmMod
         val ptr = mod.malloc(ump.size * 4)
         try {
-            ump.forEachIndexed { i, v -> mod.setValue((ptr + i * 4).toDouble(), v.toDouble(), "i32") }
+            ump.forEachIndexed { i, v -> mod.setValue(ptr + i * 4, v.toDouble(), "i32") }
             mod.uapmdEngineEnqueueUmp(handle, instanceId, ptr, ump.size, timestamp.toDouble())
         } finally { mod.free(ptr) }
     }
